Move onPlayProgressInfoUpdateListener object creation to viewHolder init
